Step_2/requirements.txt
# Default Dependencies
docling==2.31.0
google-generativeai==0.8.3
sentence-transformers>=2.2.0
lancedb==0.22.0
python-dotenv>=0.21.0
pydantic>=2.0.0

